+++ tests/integration/test_trackingclass.py
@@ -1,27 +1,12 @@
-import os
-import unittest
 from datetime import datetime
 
-from quickbooks.auth import Oauth1SessionManager
-from quickbooks.client import QuickBooks
 from quickbooks.objects.trackingclass import Class
+from tests.integration.test_base import QuickbooksTestCase
 
 
-class ClassTest(unittest.TestCase):
+class ClassTest(QuickbooksTestCase):
     def setUp(self):
-        self.session_manager = Oauth1SessionManager(
-            sandbox=True,
-            consumer_key=os.environ.get('CONSUMER_KEY'),
-            consumer_secret=os.environ.get('CONSUMER_SECRET'),
-            access_token=os.environ.get('ACCESS_TOKEN'),
-            access_token_secret=os.environ.get('ACCESS_TOKEN_SECRET'),
-        )
-
-        self.qb_client = QuickBooks(
-            session_manager=self.session_manager,
-            sandbox=True,
-            company_id=os.environ.get('COMPANY_ID')
-        )
+        super(ClassTest, self).setUp()
 
         self.name = "Test Class {0}".format(datetime.now().strftime('%d%H%M'))