+++ tests/integration/test_transfer.py
@@ -1,28 +1,15 @@
-import os
-import unittest
+import time
 from datetime import datetime
 
-from quickbooks.auth import Oauth1SessionManager
-from quickbooks.client import QuickBooks
 from quickbooks.objects.account import Account
 from quickbooks.objects.transfer import Transfer
+from tests.integration.test_base import QuickbooksTestCase
 
 
-class TransferTest(unittest.TestCase):
+class TransferTest(QuickbooksTestCase):
     def setUp(self):
-        self.session_manager = Oauth1SessionManager(
-            sandbox=True,
-            consumer_key=os.environ.get('CONSUMER_KEY'),
-            consumer_secret=os.environ.get('CONSUMER_SECRET'),
-            access_token=os.environ.get('ACCESS_TOKEN'),
-            access_token_secret=os.environ.get('ACCESS_TOKEN_SECRET'),
-        )
-
-        self.qb_client = QuickBooks(
-            session_manager=self.session_manager,
-            sandbox=True,
-            company_id=os.environ.get('COMPANY_ID')
-        )
+        time.sleep(3)  # Used to prevent error code 3001 - The request limit was reached.
+        super(TransferTest, self).setUp()
 
         self.account_number = datetime.now().strftime('%d%H%M')
         self.name = "Test Transfer {0}".format(self.account_number)