+++ tests/unit/test_mixins.py
@@ -2,7 +2,9 @@ import os
 import unittest
 
 from quickbooks.objects import Bill, Invoice
-from quickbooks.auth import Oauth1SessionManager
+
+from intuitlib.client import AuthClient
+from quickbooks.client import QuickBooks
 
 try:
     from mock import patch
@@ -130,18 +132,18 @@ class ToDictMixinTest(unittest.TestCase):
 
 class ListMixinTest(unittest.TestCase):
     def setUp(self):
-        self.session_manager = Oauth1SessionManager(
-            sandbox=True,
-            consumer_key="update_consumer_key",
-            consumer_secret="update_consumer_secret",
-            access_token="update_access_token",
-            access_token_secret="update_access_token_secret",
+        self.auth_client = AuthClient(
+            client_id=os.environ.get('CLIENT_ID'),
+            client_secret=os.environ.get('CLIENT_SECRET'),
+            environment='sandbox',
+            redirect_uri='http://localhost:8000/callback',
         )
 
-        self.qb_client = client.QuickBooks(
-            session_manager=self.session_manager,
+        self.qb_client = QuickBooks(
+            auth_client=self.auth_client,
+            refresh_token=os.environ.get('REFRESH_TOKEN'),
+            company_id=os.environ.get('COMPANY_ID'),
             sandbox=True,
-            company_id="COMPANY_ID"
         )
 
     @patch('quickbooks.mixins.ListMixin.where')